requirements.txt
pygame==2.5.2
pycryptodome==3.19.0